requirements.txt
streamlit>=1.33.0
openai>=1.0.0
chromadb>=0.4.0
jieba>=0.42.1
rank-bm25>=0.2.2
PyPDF2>=3.0.0
PyMuPDF>=1.23.5
python-pptx>=0.6.21
docx2txt>=0.8
google-genai>=0.1.0
tqdm>=4.66.0
numpy>=1.24.0
